HelpWindowTest: Move to systemtests
As AddressBookGuiTest and AddressBookSystemTest are very similar, it’s
not necessary to have both of them.

As part of removing AddressBookGuiTest, let’s move HelpWindowTest to
systemtests. Also, let’s rename it to HelpCommandTest to better reflect
what it is testing for.
