+++ src/test/java/systemtests/HelpCommandSystemTest.java
@@ -1,18 +1,24 @@
-package guitests;
+package systemtests;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import guitests.GuiRobot;
 import guitests.guihandles.HelpWindowHandle;
 import seedu.address.logic.commands.HelpCommand;
 
-public class HelpWindowTest extends AddressBookGuiTest {
+/**
+ * TODO: This test is incomplete as it is missing test cases.
+ */
+public class HelpCommandSystemTest extends AddressBookSystemTest {
     private static final String ERROR_MESSAGE = "ATTENTION!!!! : On some computers, this test may fail when run on "
             + "non-headless mode as FxRobot#clickOn(Node, MouseButton...) clicks on the wrong location. We suspect "
             + "that this is a bug with TestFX library that we are using. If this test fails, you have to run your "
             + "tests on headless mode. See UsingGradle.adoc on how to do so.";
+
+    private final GuiRobot guiRobot = new GuiRobot();
 
     @Test
     public void openHelpWindow() {
@@ -38,7 +44,7 @@ public class HelpWindowTest extends AddressBookGuiTest {
         assertHelpWindowOpen();
 
         //use command box
-        runCommand(HelpCommand.COMMAND_WORD);
+        executeCommand(HelpCommand.COMMAND_WORD);
         assertHelpWindowOpen();
     }
 
@@ -50,7 +56,7 @@ public class HelpWindowTest extends AddressBookGuiTest {
         guiRobot.pauseForHuman();
 
         new HelpWindowHandle(guiRobot.getStage(HelpWindowHandle.HELP_WINDOW_TITLE)).close();
-        mainWindowHandle.focus();
+        getMainWindowHandle().focus();
     }
 
     /**
